added methods createUser and deleteUser

+++ src/main/java/com/example/practice/entity/Category.java
@@ -9,10 +9,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Category {
 
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private String image;
-    private Integer parentCategoryId;
+    private Long parentCategoryId;
 
 }
